refactor(chat): external-provider 600줄 가드 분할 (683→554)
결정적 첨부 후처리(external-deterministic-append)와 접근불가 모델
영속화(external-model-availability)를 모듈로 추출. 순수 추출 —
export 시그니처·호출부 무변경.
